added comments, added same id check to delete

+++ backend/api/routes/user.routes.js
@@ -6,6 +6,7 @@ const bcrypt = require('bcrypt');
 const router = express.Router();
 const { checkSession } = require('../middleware.js')
 
+// Login the user, using either email or username, and password
 router.post('/login/', async (req, res) => {
     try {
         const { identifier, password } = req.body;
@@ -22,6 +23,7 @@ router.post('/login/', async (req, res) => {
     }
 });
 
+// Logout the user by destroying the session
 router.get('/logout/', checkSession, async (req, res) => {
     req.session.destroy(function (err) {
         if (err) {
@@ -35,6 +37,7 @@ router.get('/logout/', checkSession, async (req, res) => {
     });
 });
 
+// Returns information (excluding password hash) about the user matching email or password
 router.get('/lookup/:email_username', async (req, res) => {
     const identifier = req.params.email_username;
     try {
@@ -47,6 +50,7 @@ router.get('/lookup/:email_username', async (req, res) => {
     }
 });
 
+//Returns information (excluding password hash) about the user matching id
 router.get('/:id', async (req, res) => {
     try {
         const existingUser = await User.findById(req.params.id).select('-password');
@@ -58,6 +62,7 @@ router.get('/:id', async (req, res) => {
     }
 });
 
+//Creates new user with given information
 router.post('/signup/', async (req, res) => {
     try {
         const newUser = new User({
@@ -84,10 +89,23 @@ router.post('/signup/', async (req, res) => {
     }
 });
 
+//Deletes user if authorized (session has same userId), destroys session
 router.delete('/:id', checkSession, async (req, res) => {
     try {
+        if(req.session.userId != req.params.id) {
+            res.status(401).send("Unauthorized");
+            return;
+        }
         const deletedUser = await User.findOneAndDelete({ _id: req.params.id });
-        if (deletedUser) res.send("Deleted user.");
+        if (deletedUser) {
+            res.send("Deleted user.");
+            req.session.destroy(function (err) {
+                if (err) {
+                    // Handle error
+                    console.error("Session destruction error:", err);
+                } 
+            });
+        }
         else res.status(404).send("User not found.");
     } catch (error) {
         res.status(400).send("Error");
